v2025.4.1

+++ src/main/kotlin/gogo/gogoshop/domain/shop/root/event/TicketShopBuyEvent.kt
@@ -10,6 +10,6 @@ data class TicketShopBuyEvent(
     val shopMiniGameId: Long,
     val ticketType: TicketType,
     val shopReceiptId: Long,
-    val ticketPrice: Int,
+    val ticketPrice: Long,
     val purchaseQuantity: Int,
 )
